% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_enable_project.R
\name{enable_project}
\alias{enable_project}
\title{Enable an existing project}
\usage{
enable_project(root_folder = getwd(), first_name, last_name,
  full_path = NULL)
}
\arguments{
\item{root_folder}{Allows to change the location in which the function
creates the new project folder. dateFirst and last name together with
the degree are forming the projct folder name inside the
root_folder (see next params)}

\item{first_name}{Allows to provide the first name of the creator and
reponsible person for the project.
project (mandatory)}

\item{last_name}{Allows to provide the last name of the creator and
responsible person for the project.
project (mandatory)}

\item{full_path}{Allows you to specify the full path to a project root folder.
This also allows to createa a flexible named folder for projects.}
}
\description{
Enable an existing project
}
