% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_handle_options.R
\name{handle_options}
\alias{handle_options}
\title{Get and set options for the package}
\usage{
handle_options(...)
}
\arguments{
\item{...}{Pass in arbitrary names of options in the named list.
to query for the value. Pass in a name and a value to set
the value to your liking.}
}
\description{
Get and set options for the package
}
\examples{
\dontrun{
 # query
 agwdm::handle_options("xyz")
 # set
 agwdm::handle_options("xyz" = "value")
}

}
