% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_project_path.R
\name{project_content}
\alias{project_content}
\title{A fuction to compile project paths}
\usage{
project_content(path = NULL)
}
\arguments{
\item{path}{The path to the file or folder inside the project.}
}
\value{
a string path pointing to the file or folder
}
\description{
This function compiles a project path which is always constructued
from the root folder of the project. As the predefined folder structure
has the folder "project" on top this is taken as the root inside the
actual project name. This makes the project more self contained and
allows references to files which do not break.
}
\examples{
\dontrun{
 project_content("data/raw/iris.csv")
     "project_name/project/data/raw/iris.csv"
}

}
