% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_handle_options.R
\name{handle_options}
\alias{handle_options}
\title{Get and set options for the package}
\usage{
handle_options(...)
}
\arguments{
\item{...}{Pass in arbitrary names of options in the named list.
to query for the value. Pass in a name and a value to set
the value to your liking.}
}
\description{
the package has an own environment with a list of options.
it is storing e.g. paths to folders and files in the project
structure the mspm package generates. It will be used by the
package itself to enable a project structure awareness throughout
the mspm package and allows convenient functions for users to
reference locations and files while writing conducting their
analysis using the project template.
}
\examples{
\dontrun{
 # query
 agwdm::handle_options("xyz")
 # set
 agwdm::handle_options("xyz" = "value")
}

}
